Add basic CI tests, renovate config, and move protobufs to own directory.
