+++ protos/faucetconfrpc.proto
@@ -1,5 +1,5 @@
 // From directory with this definition, run this to regenerate python libraries:
-// python -m grpc_tools.protoc --python_out=.. --grpc_python_out=.. -I . faucetconfrpc.proto
+// python -m grpc_tools.protoc --python_out=. --grpc_python_out=. -I . faucetconfrpc.proto
 
 syntax = "proto3";
