+++ test-requirements.txt
@@ -1,0 +1,3 @@
+-r requirements.txt
+pylint==2.5.3
+pytype==2020.6.1
